% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffuse1d.R
\name{diffuse}
\alias{diffuse}
\title{Bestimme aktuelle Konzentration für Feld i}
\usage{
diffuse(i, x, delta)
}
\arguments{
\item{i}{Integer, Positionsindex}

\item{x}{Vektor aus reellen Zahler: Konzentrartion}

\item{delta}{Diffusionsfaktor}
}
\value{
reelle Zahl, Konzentration an Stelle i in nächsten
Zeitschritt
}
\description{
Bestimme aktuelle Konzentration für Feld i
}
