% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffuse2d.R
\name{diffuse2D}
\alias{diffuse2D}
\title{Modelliere Diffusion in 2d}
\usage{
diffuse2D(A, delta)
}
\arguments{
\item{A}{reelle Matrix:  Konzentrationen in x,y Koordinaten}

\item{delta}{Diffusionskoeffizient, Anteil der Abwanderung in
benachbarte Zellen}
}
\value{
reelle Matrix: Konzentrationen in x,y Koordinaten nach erfolgter
Diffusion
}
\description{
Modelliere Diffusion in 2d
}
