% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lotka-volterra.R
\name{LotVmod}
\alias{LotVmod}
\title{Modellfunktion für Lotka-Volterra-Modell}
\usage{
LotVmod(Time, State, Pars)
}
\arguments{
\item{Time}{Integer, Index der Zeit}

\item{State}{Liste mit Integern \code{State$B} = Beutepopulation und
\code{State$R} = Räuberpopulation}

\item{Pars}{Liste mit Reals \code{Pars$r} = Wachstum Beute
\code{Pars$m} = Schwund Räuber
\code{Pars$a} = Jagderfolg
\code{Pars$b} = Umwandlung Räuber}
}
\value{
Liste mit Feldern \code{dB} Veränderung der Beutepopulation und
\code{dR} Veränderung der Räuberpopulation
}
\description{
LotVmod(Time, State, Pars) gibt die Veränderungen der Popoulationen
in einem Zeitschritt wider, je nach vorigen Populationen und Parametern
}
