% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diskret_dynamische_Lachse.R
\name{model}
\alias{model}
\title{Modellfunktion zur Vorhersage einer Lachspopulation}
\usage{
model(S4, S5, E = 0.57)
}
\arguments{
\item{S4}{Population 4 Jahre vorher}

\item{S5}{Population 5 Jahre vorher}

\item{E}{Laicherfolg als Zahl zwischen 0 und 1}
}
\description{
Modellfunktion zur Vorhersage einer Lachspopulation
}
