% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diskret_dynamischer_Stausee.R
\name{modell}
\alias{modell}
\title{Berechne Abfluss eines Stausees in Abhängigkeit vom Eintrag}
\usage{
modell(s0, input, k, schritte)
}
\arguments{
\item{s0}{Menge im Speicher zu Beginn}

\item{input}{kontinuierlicher Eintrag pro periode}

\item{k}{Verhältnis Abfluss/Gesamtmenge}

\item{schritte}{Dauer der Simulation}
}
\description{
Berechne Abfluss eines Stausees in Abhängigkeit vom Eintrag
}
