% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistisches_wachstum.R
\name{Nt}
\alias{Nt}
\title{Modellfunktion für diskretes logistisches Wachstum}
\usage{
Nt(r, K, N_k, t)
}
\arguments{
\item{r}{Zahl zwischen -0.1 und 4, gibt Wachstumsrate an
(Verhältnis zusätzliche Individuen zum nächsten Zeitpunkt zu aktuellen)}

\item{K}{Zahl, gibt Carrying Capacity der Umwelt an ~ max Population}

\item{N_k}{Zahl zwischen 0 und 1, relative Anfangspopulation
im Verhältnis zur Kapazität an}

\item{t}{Zahl Endzeitpunkt der Simulation}
}
\value{
Vektor von Zahlen, hält am Index i Populationsgröße zur Zeit i
}
\description{
\code{Nt(r, K, N_k, t)} erstellt einen Vektor der Populationsgröße zum
Zeitpunkt i am Index i.
}
