% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random1d.R
\name{random1d}
\alias{random1d}
\title{Simuliere einen Münzwurf}
\usage{
random1d(p)
}
\arguments{
\item{p}{reelle Zahl, Wahrscheinlichkeit für Bild}
}
\value{
Integer  +1 (zu p * 100 \%) oder −1 (zu (1−p) \* 100 \%
}
\description{
Simuliere einen Münzwurf
}
