% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random1d.R
\name{random1d_c}
\alias{random1d_c}
\title{Simuliere einen kontinuierlichen 1d - Schritt}
\usage{
random1d_c(mean, sd)
}
\arguments{
\item{mean}{reelle Zahl, Erwartungswert}

\item{sd}{reelle Zahl, Standardabweichung}
}
\value{
reelle Zahl, zufälligen Wert aus der Normalverteilung
}
\description{
Simuliere einen kontinuierlichen 1d - Schritt
}
