% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random2d.R
\name{random2d}
\alias{random2d}
\title{Simuliere einen RandomWalk in 2D}
\usage{
random2d(x, y, a, speed, rho)
}
\arguments{
\item{x}{reelle Zahl, aktuelle X-Koordinate}

\item{y}{reelle Zahl, aktuelle Y-Koordinate}

\item{a}{reelle Zahl 0 ≤ a ≤ 2π: Richtungswinkel der vorigen Bewegung}

\item{speed}{reelle Zahl, Erwartungswert Geschwindigkeit (Normalverteilung)}

\item{rho}{reele Zahl 0 ≤ rho ≤ 1: Wahrscheinlichkeit des "Beharrens" auf
voriger Bewegungsrichtung}
}
\value{
Vektor von 3 reelen Zahlen: x, y Koordinaten und
eingeschlagene Richtung
}
\description{
Simuliere einen RandomWalk in 2D
}
