% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistisches_wachstum.R
\name{Nt}
\alias{Nt}
\title{Modellfunktion für diskretes logistisches Wachstum}
\usage{
Nt(r, K, N, t)
}
\arguments{
\item{r}{Zahl zwischen 0 und 1, gibt Wachstumsrate an}

\item{K}{Zahl, gibt Carrying Capacity der Umwelt an ~ max Population}

\item{N}{Array von Zahlen, hält am Index 1 die Populationsgröße zur Zeit 1}

\item{t}{Zahl Endzeitpunkt der Simulation}
}
\value{
Array von Zahlen, hält am Index t Populationsgröße zur Zeit t
}
\description{
Modellfunktion für diskretes logistisches Wachstum
}
